renaming and updating manager
